fea: nota postoperatoria correccion de mensajes de error y del request

+++ app/Http/Requests/TrasladoRequest.php
@@ -20,7 +20,7 @@ class TrasladoRequest extends FormRequest
             'fr'   => ['required', 'numeric'],
             'temp' => ['required', 'numeric'],
             'peso' => ['required', 'numeric'],
-            'talla'=> ['required', 'numeric'],
+            'talla'=> ['required', 'integer'],
 
             // ---- Consulta / Exploración (Todos obligatorios ahora) ----
             'resultado_estudios'               => ['required', 'string'],
